// src/user/schemas/user.entity.ts
//  Generated by entity-cli

import * as mongoose from "mongoose";
import { role } from "../role.enum";

export const schemaName = "User";

export interface UserInterface extends mongoose.Document {
  nom: string;
  prenom: string;
  description: string;
  username: string;
  password: string;
  imagePath: string;
  role: role;
}

export const User = new mongoose.Schema(
  {
    nom: {
      type: String,
      require: true,
    },
    prenom: {
      type: String,
      require: true,
    },
    description: {
      type: String,
      require: true,
    },
    username: {
      type: String,
      require: true,
    },
    password: {
      type: String,
      require: true,
    },
    imagePath: {
      type: String,
      require: false,
      default: "null",
    },
    role: {
      type: String,
      enum: role,
      require: true,
    },
  },
  { timestamps: true },
);